Backend with authentication ready

+++ FitnessApp.API/Services/RecurringJobsService.cs
@@ -30,40 +30,51 @@ namespace FitnessApp.API.Services
 
         public void ClearTheDailyIntakeMacrosAndMoveItToHistory()
         {
-            var dailySummedMacros = _macrosRepository.GetDailyMealMacrosSummed();
-            var dailyMacrosIntakeForHistoryDto = new DailyMacrosIntakeHistoryDto();
-            if (dailySummedMacros == null)
+            var dailySummedMacrosList = _macrosRepository.GetDailyMealMacrosSummedForAllUsers();
+            if(dailySummedMacrosList.Count != 0)
             {
-                dailyMacrosIntakeForHistoryDto = new DailyMacrosIntakeHistoryDto()
+                foreach (var dailySummedMacros in dailySummedMacrosList)
                 {
-                    UId = Guid.NewGuid(),
-                    CreatedOn = DateTime.Now,
-                    Protein = 0,
-                    Carbs = 0,
-                    Fats = 0
+                    var dailyMacrosIntakeForHistoryDto = new DailyMacrosIntakeHistoryDto();
+                    if (dailySummedMacros == null)
+                    {
+                        dailyMacrosIntakeForHistoryDto = new DailyMacrosIntakeHistoryDto()
+                        {
+                            UId = Guid.NewGuid(),
+                            CreatedOn = DateTime.Now,
+                            Protein = 0,
+                            Carbs = 0,
+                            Fats = 0,
+                            UserFk = dailySummedMacros.UserFk
 
-                };
-               
-            } else
-            {
-                dailyMacrosIntakeForHistoryDto = new DailyMacrosIntakeHistoryDto()
-                {
-                    UId = Guid.NewGuid(),
-                    CreatedOn = DateTime.Now,
-                    Protein = dailySummedMacros.Protein,
-                    Carbs = dailySummedMacros.Carbs,
-                    Fats = dailySummedMacros.Fats
-                };
+                        };
+
+
+                    }
+                    else
+                    {
+                        dailyMacrosIntakeForHistoryDto = new DailyMacrosIntakeHistoryDto()
+                        {
+                            UId = Guid.NewGuid(),
+                            CreatedOn = DateTime.Now,
+                            Protein = dailySummedMacros.Protein,
+                            Carbs = dailySummedMacros.Carbs,
+                            Fats = dailySummedMacros.Fats,
+                            UserFk = dailySummedMacros.UserFk
+
+                        };
+                    }
+
+                    var finalDailyMacrosIntakeForHistory = _mapper.Map<Entities.DailyMacroIntakeHistory>(dailyMacrosIntakeForHistoryDto);
+
+                    _macrosRepository.AddDailyIntakeToHistory(finalDailyMacrosIntakeForHistory);
+
+                    _macrosRepository.Save();
+                }
+                _macrosRepository.DeleteAllMealMacros();
+                _macrosRepository.Save();
             }
-
-             
-
-            var finalDailyMacrosIntakeForHistory = _mapper.Map<Entities.DailyMacroIntakeHistory>(dailyMacrosIntakeForHistoryDto);
-
-            _macrosRepository.AddDailyIntakeToHistory(finalDailyMacrosIntakeForHistory);
-            _macrosRepository.DeleteAllMealMacros();
-            _macrosRepository.Save();
-
+            
         }
